Nick Williams and Merijn Broeren at MSDW did the impossible and managed to make POE::Wheel::Run pass actual tests on ActivePerl. Damn, they're GOOD!  This round of patches isn't 100% perfect (there's an issue with the Wheel under POE::Loop::Tk), but I'm confident it'll be CPAN-worthy shortly.

+++ lib/POE/Wheel/Run.pm
@@ -18,7 +18,6 @@ use POE qw( Wheel Pipe::TwoWay Pipe::OneWay Driver::SysRW Filter::Line );
 
 BEGIN {
   die "$^O does not support fork()\n" if $^O eq 'MacOS';
-  die "$^O does not fully support fork+exec\n" if $^O eq 'MSWin32';
 
   local $SIG{'__DIE__'} = 'DEFAULT';
   eval    { require IO::Pty; };
@@ -26,6 +25,16 @@ BEGIN {
   else {
     IO::Pty->import();
     eval 'sub PTY_AVAILABLE () { 1 }';
+  }
+
+  if (POE::Kernel::RUNNING_IN_HELL) {
+      eval    { require Win32::Console; };
+      if ($@) { die "Win32::Console failed to load:\n$@" }
+      else    { Win32::Console->import(); };
+
+      eval    { require Win32API::File; };
+      if ($@) { die "Win32API::File but failed to load:\n$@" }
+      else    { Win32API::File->import( qw(FdGetOsFHandle) ); };
   }
 
   # How else can I get them out?!
@@ -387,41 +396,66 @@ sub new {
     close $stdin_write;
     close $stdout_read;
     close $stderr_read if defined $stderr_read;
+    
+    # Need to close on Win32 because std handles aren't dup'ed, no
+    # harm elsewhere.  Close STDERR later to not influence possible
+    # die.
+    close STDIN;
+    close STDOUT;
 
     # Redirect STDIN from the read end of the stdin pipe.
     open( STDIN, "<&" . fileno($stdin_read) )
       or die "can't redirect STDIN in child pid $$: $!";
-    fileno(STDIN) == STDIN_FILENO
-      or die "child's STDIN filehandle does not have fileno of ".STDIN_FILENO;
+    POE::Kernel::RUNNING_IN_HELL or fileno(STDIN) == STDIN_FILENO or die (
+      "child's STDIN fileno == ", fileno(STDIN),
+      " but ", STDIN_FILENO, " was expected"
+    );
 
     # Redirect STDOUT to the write end of the stdout pipe.
+    # The STDOUT_FILENO check snuck in on a patch.  I'm not sure why
+    # we care what the file descriptor is.
     open( STDOUT, ">&" . fileno($stdout_write) )
       or die "can't redirect stdout in child pid $$: $!";
-    fileno(STDOUT) == STDOUT_FILENO
-      or die "child's STDOUT filehandle does not have fileno of ".STDOUT_FILENO;
+    POE::Kernel::RUNNING_IN_HELL or fileno(STDOUT) == STDOUT_FILENO or die (
+      "child's STDOUT fileno == ", fileno(STDOUT),
+      " but ", STDOUT_FILENO, " was expected"
+    );
+
+    # Need to close on Win32 because std handles aren't dup'ed, no harm elsewhere
+    close STDERR;
 
     # Redirect STDERR to the write end of the stderr pipe.  If the
     # stderr pipe's undef, then we use STDOUT.
+    # The STDERR_FILENO check snuck in on a patch.  I'm not sure why
+    # we care what the file descriptor is.
     open( STDERR, ">&" . fileno($stderr_write) )
       or die "can't redirect stderr in child: $!";
-    fileno(STDERR) == STDERR_FILENO
-      or die "child's STDERR filehandle does not have fileno of ".STDERR_FILENO;
+    POE::Kernel::RUNNING_IN_HELL or fileno(STDERR) == STDERR_FILENO or die (
+      "child's STDERR fileno == ", fileno(STDERR),
+      " but ", STDERR_FILENO, " was expected"
+    );
 
     # Make STDOUT and/or STDERR auto-flush.
     select STDERR;  $| = 1;
     select STDOUT;  $| = 1;
 
     # Tell the parent that the stdio has been set up.
-    close $sem_pipe_read unless $^O eq 'MSWin32';
+    close $sem_pipe_read;
     print $sem_pipe_write "go\n";
-    close $sem_pipe_write unless $^O eq 'MSWin32';
+    close $sem_pipe_write;
 
-    # Exec the program depending on its form.
-    if (ref($program) eq 'ARRAY') {
-      exec(@$program, @$prog_args)
-        or die "can't exec (@$program) in child pid $$: $!";
+    if (POE::Kernel::RUNNING_IN_HELL)  {
+	# The Win32 pseudo fork sets up the std handles in the child based on the true win32 handles
+	# For the exec these get remembered, so manipulation of STDIN/OUT/ERR is not enough. Only 
+	# necessary for the exec, as Perl CODE subroutine goes through 0/1/2 which are correct. 
+	# But ofcourse that coderef might invoke exec, so better do it regardless.
+	# HACK: Using Win32::Console as nothing else exposes SetStdHandle
+	Win32::Console::_SetStdHandle(STD_INPUT_HANDLE(),  FdGetOsFHandle(fileno($stdin_read)));
+	Win32::Console::_SetStdHandle(STD_OUTPUT_HANDLE(), FdGetOsFHandle(fileno($stdout_write)));
+	Win32::Console::_SetStdHandle(STD_ERROR_HANDLE(),  FdGetOsFHandle(fileno($stderr_write)));
     }
-    elsif (ref($program) eq 'CODE') {
+    # Exec the program depending on its form.
+    if (ref($program) eq 'CODE') {
 
       # Close any close-on-exec file descriptors.
       if ($close_on_call) {
@@ -436,16 +470,23 @@ sub new {
 
       # Try to exit without triggering END or object destructors.
       # Give up with a plain exit if we must.
-      eval { POSIX::_exit(0);  };
-      eval { kill KILL => $$;  };
-      eval { exec("$^X -e 0"); };
+      # On win32 cannot _exit as it will kill *all* threads, meaning parent too
+      unless (POE::Kernel::RUNNING_IN_HELL) {
+	  eval { POSIX::_exit(0);  };
+	  eval { kill KILL => $$;  };
+	  eval { exec("$^X -e 0"); };
+      };
       exit(0);
+    } else {
+	if (ref($program) eq 'ARRAY') {
+	  exec(@$program, @$prog_args)
+	    or die "can't exec (@$program) in child pid $$: $!";
+	}
+	else {
+	  exec(join(" ", $program, @$prog_args))
+	    or die "can't exec ($program) in child pid $$: $!";
+	}
     }
-    else {
-      exec(join(" ", $program, @$prog_args))
-        or die "can't exec ($program) in child pid $$: $!";
-    }
-
     die "insanity check passed";
   }
